Merged to Dev | Sprint 5 | Feature Algorithm

See merge request courses/2022-winter/csci-5308/group20!81

+++ src/main/java/com/scheduler/app/model/request/EmployeeCreationRequest.java
@@ -8,6 +8,9 @@ import lombok.Setter;
 
 import java.time.LocalDate;
 
+/**
+ * The type - Request Object for Employee Creation.
+ */
 @Getter
 @Setter
 @NoArgsConstructor
@@ -20,7 +23,7 @@ public class EmployeeCreationRequest {
     private String loginPassword = "XYZ";
     private String firstName;
     private String lastName;
-    private Integer phoneNumber;
+    private Long phoneNumber;
     private Integer roleId;
     private byte[] photo;
     private String sinNumber;
